platforms/gcp/requirements.txt
# GCP Cloud Functions
functions-framework>=3.0.0
flask>=2.0.0

# LangChain framework
langchain>=0.2.0
langchain-core>=0.2.0
langgraph>=0.2.0

# Cloud LLM providers
langchain-openai>=0.1.0              # Azure OpenAI (if needed)
langchain-google-vertexai>=1.0.0     # GCP Vertex AI (Gemini)
langchain-aws>=0.1.0                 # AWS Bedrock (if needed)

# Utilities
python-dotenv>=1.0.0

# Document processing
pypdf>=3.0.0
openpyxl>=3.1.0
python-docx>=1.0.0

# ==============================================================================
# OCR Providers
# ==============================================================================

# GCP Document AI (recommended for GCP)
google-cloud-documentai>=2.0.0

# Azure Document Intelligence (optional)
# azure-ai-documentintelligence>=1.0.0
# azure-core>=1.30.0

# Tesseract OCR (optional - local OSS)
# pytesseract>=0.3.10
# pdf2image>=1.16.0
# Pillow>=10.0.0
